requirements.txt
black
flake8
isort<5
pytest
click
sphinx
bloomberg.ai.remoteio
bloomberg.ds.katiecli
bloomberg.ds.maestro.client
sentence-transformers
botocore
boto3
nltk
munch
torch
pandas
pillow
tqdm
IPython
typo
spacy